services/config_backup_worker/requirements.txt
netmiko>=4.3.0
psycopg2-binary>=2.9.9
schedule>=1.2.0
pika>=1.3.2
